Move the module into core to sit alongside the rest of the core-sourced logic
