+++ interface/modules/zend_modules/config/application.config.php
@@ -20,7 +20,6 @@ $core_modules = [
     'Installer', // Handles the dynamic adding / removing of modules in the system.
     'Acl', // Handles all of the permission checks in the system.
     'FHIR', // Handles FHIR mapped uuid population and other FHIR utility functions
-    'CodeTypes', // Handles CodeType mappings and anything else to do with the system of dealing with code types
     'PatientFlowBoard', // Handle any functionality needed for the patient flow board
 ];
