--- interface/modules/zend_modules/module/CodeTypes/config/module.config.php
@@ -1,24 +1,0 @@
-<?php
-
-/**
- * Handles class, routing, views, and other configuration properties for the module.
- *
- * @package   OpenEMR
- * @link      https://www.open-emr.org
- * @author    Stephen Nielson <stephen@nielson.org>
- * @copyright Copyright (c) 2021 Stephen Nielson <stephen@nielson.org>
- * @license   https://github.com/openemr/openemr/blob/master/LICENSE GNU General Public License 3
- */
-
-namespace OpenEMR\ZendModules\CodeTypes;
-
-use Laminas\ServiceManager\Factory\InvokableFactory;
-use OpenEMR\ZendModules\CodeTypes\Listener\CodeTypeEventsSubscriber;
-
-return [
-    'service_manager' => [
-        'factories' => [
-            CodeTypeEventsSubscriber::class => InvokableFactory::class
-        ]
-    ]
-];
